knowledge_assistant/api/serializers.py
from rest_framework import serializers
from .models import UploadedDocument

class UploadedDocumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = UploadedDocument
        fields = ['id', 'file', 'uploaded_at', 'processed']